+++ DawnLib.Dusk/src/API/Definitions/EntityReplacement/EntityReplacementConfig.cs
@@ -7,4 +7,5 @@ public class EntityReplacementConfig
     public ConfigEntry<string>? MoonSpawnWeights;
     public ConfigEntry<string>? InteriorSpawnWeights;
     public ConfigEntry<string>? WeatherSpawnWeights;
+    public ConfigEntry<bool>? DisableDateCheck;
 }