+++ DawnLib.Dusk/src/API/DuskModConstants.cs
@@ -4,8 +4,8 @@ static class DuskModConstants
     internal const string MenuName = "DawnLib";
     internal const string Definitions = $"{MenuName}/Definitions";
     internal const string Achievements = $"{MenuName}/Definitions/Achievements";
-    internal const string EntityReplacements = $"{MenuName}/Definitions/Entity Replacements";
     internal const string TerminalPredicates = $"{MenuName}/Terminal Predicates";
+    internal const string DuskPredicates = $"{MenuName}/Dusk Predicates";
     internal const string PricingStrategies = $"{MenuName}/Pricing Strategies";
 
     internal const int PredicateOrder = -50;